+++ foxes/output/__init__.py
@@ -1,6 +1,7 @@
 """
 Output tools and functions.
 """
+from .round import round_defaults
 from .output import Output
 from .farm_layout import FarmLayoutOutput
 from .farm_results_eval import FarmResultsEval
@@ -9,5 +10,6 @@ from .results_writer import ResultsWriter
 from .state_turbine_map import StateTurbineMap
 from .turbine_type_curves import TurbineTypeCurves
 from .animation import Animator
+from .calc_points import PointCalculator
 
 from .flow_plots_2d import FlowPlots2D, SeqFlowAnimationPlugin